+++ rollup.config.js
@@ -6,8 +6,6 @@ const pkg = require('./package.json');
 
 export default {
   input: 'src/index.js',
-  sourcemap: true,
-  exports: 'named',
   plugins: [
     commonjs(),
     resolve(),
@@ -15,15 +13,24 @@ export default {
       exclude: 'node_modules/**',
       babelrc: false,
       presets: [
-        ['env', {
-          modules: false,
-        }],
+        [
+          'env',
+          {
+            modules: false,
+          },
+        ],
       ],
       plugins: ['external-helpers'],
     }),
   ],
   output: [
-    { file: pkg.main, format: 'umd', name: 'rest-client-sdk', exports: 'named' },
-    { file: pkg.module, format: 'es' },
+    {
+      file: pkg.main,
+      format: 'umd',
+      name: 'rest-client-sdk',
+      exports: 'named',
+      sourcemap: true,
+    },
+    { file: pkg.module, format: 'es', sourcemap: true },
   ],
 };