+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.0.0-rc.1
+
 ### Added
 
 * [Might break] Custom serializer can: